fixes, helper methods, more font and text-path related tests

+++ raphael4gwt/src/org/sgx/raphael4gwt/raphael/Raphael.java
@@ -202,7 +202,7 @@ public static native int deg(int deg)/*-{
 //}
 
 
-public static native Point createPoint(int x, int y)/*-{
+public static native Point createPoint(double x, double y)/*-{
 	return {"x": x, "y": y};
 }-*/;
 public static native Rectangle createRectangle(int x, int y, int width, int height)/*-{
